refactor: add new validations

+++ src/controllers/subprocess-controllers.ts
@@ -4,13 +4,30 @@ import {
   PROCESS_NOT_FOUND_ERROR
 } from "../domain/error/erros";
 import { SubprocessServices } from "../services/subprocess-service";
+import { Validators } from "../utils/utils";
+import {
+  badrequestError,
+  notFoundError,
+  ok,
+  responseError
+} from "./adapters/handlers";
 
 export class SubprocessController {
-  constructor(private readonly subprocessService: SubprocessServices) {}
+  constructor(
+    private readonly subprocessService: SubprocessServices,
+    private readonly validators: Validators
+  ) {}
 
   createSubprocess = async (req: Request, res: Response) => {
     try {
       const { departamentId, processId } = req.params;
+
+      [processId, departamentId].forEach((param) => {
+        if (!this.validators.isValidObjectId(param)) {
+          return badrequestError(res, param);
+        }
+      });
+
       const createSubprocessDto = req.body;
 
       const processCreated = await this.subprocessService.createSubprocess({
@@ -21,14 +38,23 @@ export class SubprocessController {
 
       return res.json(processCreated);
     } catch (error) {
-      console.log(error);
+      return responseError(res, error);
     }
   };
 
   findSubprocess = async (req: Request, res: Response) => {
     try {
       const { id } = req.params;
-      const { processId, departamentId } = req.query;
+      const { processId, departamentId } = req.query as unknown as {
+        processId: string;
+        departamentId: string;
+      };
+
+      [id, processId, departamentId].forEach((param) => {
+        if (!this.validators.isValidObjectId(param)) {
+          return badrequestError(res, param);
+        }
+      });
 
       const processFound = await this.subprocessService.findOneSubprocess(
         id,
@@ -36,40 +62,52 @@ export class SubprocessController {
         departamentId as string
       );
 
-      return processFound != null
-        ? res.json(processFound)
-        : res
-            .status(404)
-            .json(
-              `${DEPARTAMENT_NOT_FOUND_ERROR.toLocaleLowerCase()} or ${PROCESS_NOT_FOUND_ERROR.toLocaleLowerCase()}`
-            );
+      if (!processFound) {
+        return notFoundError(
+          res,
+          `${DEPARTAMENT_NOT_FOUND_ERROR.toLocaleLowerCase()} or ${PROCESS_NOT_FOUND_ERROR.toLocaleLowerCase()}`
+        );
+      }
+
+      return ok(res, processFound);
     } catch (error) {
-      console.log(error);
+      return responseError(res, error);
     }
   };
 
   updateSubprocess = async (req: Request, res: Response) => {
     try {
       const { id } = req.params;
-      const { processId, departamentId } = req.query;
+      const { processId, departamentId } = req.query as unknown as {
+        processId: string;
+        departamentId: string;
+      };
+
+      [id, processId, departamentId].forEach((param) => {
+        if (!this.validators.isValidObjectId(param)) {
+          return badrequestError(res, param);
+        }
+      });
+
       const updatePayload = req.body;
 
-      const processFound = await this.subprocessService.updateSubprocess(
+      const subprocessUpdated = await this.subprocessService.updateSubprocess(
         id,
         processId as string,
         departamentId as string,
         updatePayload
       );
 
-      return processFound != null
-        ? res.json(processFound)
-        : res
-            .status(404)
-            .json(
-              `${DEPARTAMENT_NOT_FOUND_ERROR.toLocaleLowerCase()} or ${PROCESS_NOT_FOUND_ERROR.toLocaleLowerCase()}`
-            );
+      if (!subprocessUpdated) {
+        return notFoundError(
+          res,
+          `${DEPARTAMENT_NOT_FOUND_ERROR.toLocaleLowerCase()} or ${PROCESS_NOT_FOUND_ERROR.toLocaleLowerCase()}`
+        );
+      }
+
+      return ok(res, subprocessUpdated);
     } catch (error) {
-      console.log(error);
+      return responseError(res, error);
     }
   };
 
@@ -77,11 +115,16 @@ export class SubprocessController {
     try {
       const { id } = req.params;
 
+      if (!this.validators.isValidObjectId(id)) {
+        return badrequestError(res, id);
+      }
+
+
       await this.subprocessService.deleteSubprocess(id);
 
-      return res.json();
+      return ok(res, { message: "deleted with success" });
     } catch (error) {
-      console.log(error);
+      return responseError(res, error);
     }
   };
 
@@ -89,12 +132,16 @@ export class SubprocessController {
     try {
       const { subprocessId } = req.params;
 
+      if (!this.validators.isValidObjectId(subprocessId)) {
+        return badrequestError(res, subprocessId);
+      }
+
       const processFound = await this.subprocessService.findBySubprocessId(
         subprocessId
       );
-      return res.json(processFound);
+      return ok(res, processFound);
     } catch (error) {
-      console.log(error);
+      return responseError(res, error);
     }
   };
 }
